Add checks for 'Lincoln' and 'UNL' in brand mentions
Recent discussions have considered removing the "Lincoln"
from the branding of the University altogether. We opted not to
fully do this, but instead to advise users to try to avoid
'UNL' and "University of Nebraska-Lincoln" when referring to
this campus. This commit adds a check for this in the UNL metric,
(no points taken off), a notification for the user, tests to check
the function, and a page to test on.